changed speed to a multiplier

+++ UPDATED GROUP PROJECT/SimulationWorld.java
@@ -44,7 +44,7 @@ public class SimulationWorld extends World{
     private static final int FINAL_WAVE = 20;
 
     // Constructor for the world, initializes objects
-    public SimulationWorld(String playerModel, int maxLives, int speed, double difficulty){
+    public SimulationWorld(String playerModel, int maxLives, double speed, double difficulty){
         super(1024, 576, 1, false); // Create world with size 1024x576 pixels
 
         // Initialize game variables
